[DP-46] fix: MobileData 엔티티 필드 수정(데이터 양 타입 float로 수정, DataSellingUnit 제거) #36

+++ src/main/java/com/dapanda/product/dto/request/MobileDataCursorRequest.java
@@ -5,10 +5,10 @@ import lombok.Getter;
 @Getter
 public class MobileDataCursorRequest extends ProductCursorRequest {
 
-	private final Integer dataAmount;
+	private final Float dataAmount;
 
 	public MobileDataCursorRequest(Long cursorId, Integer size,
-			String productSortOption, Integer dataAmount) {
+			String productSortOption, Float dataAmount) {
 
 		super(cursorId, size, productSortOption);
 		this.dataAmount = dataAmount;
